Adjust app colors
