+++ GithubpullsApp/app/src/main/java/br/com/concrete/ghpulls/ui/search/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package br.com.concrete.ghpulls.ui.notifications
+package br.com.concrete.ghpulls.ui.search
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
